Copy button added
Added a copy PCN button to FBudggetMain.
Added saving functionality to FBudgetPNCAddition and CBBudgetPCN using
existing functions and copy and paste with slight modification of others
for Save and Edit.

When Copy PCN button is clicked in the UI, the current row is copied
into a new row in the database. This prevents issues with modifying the
save button. After the copy is initially saved, it works as a normal
edit.
Tested to make sure the line number increments.
Hours and Expenses copy.
Line numbers work when any item is chosen for copy. The next addition
has the correct number.
Descriptions etc. copy.
When coping an approved PNC, the new copy is not automatically approved.

-shawn@aegisresearchlabs.com