PCNNumber Edit
Added text box for updating PNC number to FBudgetPCNAddition Designer
following the PCNTitle text box as an example.
Verified that this does update the number. Does not update the main
screen with the new PCN Number.
Added PCN Number to the main UI on line 2338 of FBudgetMain
In FBudgetPCNAddition corrected the issue of the Next PCN number not
showing in the box for the newly added PCN by adding txtPCNNumber.Text =
moPCN.PCNNumber;  to the SaveNewPCN method. Also caused this to
increment in the CopyPCN with the same line of code above.
Tested adding PCN verified next number shows up and saves. Testing
coping PCN verified next PCN number appears in the editor and saves.
Verified that editing the PCN does save and update on the main screen.
-shawn@aegisresearchlabs.com